Move Cypress into dedicated test directory
