+++ src/main/aws-exports.js
