+++ src/main/utilities/loggingConfig.js
